move download all topic to button list
