+++ client/src/containers/Topic/Topic/Topic.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import Header from '../../Header';
+import Dropdown from 'react-bootstrap/Dropdown';
 import TopicData from './TopicData';
 import TopicPartitions from './TopicPartitions';
 import TopicGroups from './TopicGroups';
@@ -25,7 +26,10 @@ class Topic extends Root {
     roles: JSON.parse(sessionStorage.getItem('roles')),
     topicInternal: false,
     configs: [],
-    isAllTopicDataSelected: false
+    isAllTopicDataSelected: false,
+    downloadFormat: 'Select',
+    downloadOptions: ['Select', 'csv', 'json'],
+    messages: []
   };
 
   tabs = ['data', 'partitions', 'groups', 'configs', 'acls', 'logs'];
@@ -75,9 +79,50 @@ class Topic extends Root {
     });
   }
 
-  _handleSelectAllCheckboxChange = isAllTopicDataSelected => {
-    this.setState({ isAllTopicDataSelected: isAllTopicDataSelected });
+  _handleSelectAllCheckboxChange = (isAllTopicDataSelected, messages) => {
+    this.setState({ isAllTopicDataSelected: isAllTopicDataSelected, messages: messages });
   };
+
+  _renderDownloadFormat = isChecked => {
+    const { downloadOptions } = this.state;
+
+    let renderedOptions = [];
+    for (let option of downloadOptions) {
+      renderedOptions.push(
+        <Dropdown.Item
+          key={option}
+          se
+          disabled={isChecked === false}
+          onClick={() =>
+            this.setState({ downloadFormat: option }, () => {
+              this._handleDownloadAll(option);
+            })
+          }
+        >
+          <i className="fa fa-fw pull-left" aria-hidden="true" /> {option}
+        </Dropdown.Item>
+      );
+    }
+    return renderedOptions;
+  };
+
+  _handleDownloadAll(option) {
+    let messages = this.state.messages;
+    if (this.state.isAllTopicDataSelected && option !== 'Select') {
+      let allData = [];
+      messages.map(tableData => {
+        allData.push(tableData.value);
+        allData.push('\n');
+      });
+      const a = document.createElement('a');
+      const type = 'text/' + option;
+      a.href = URL.createObjectURL(new Blob([allData], { type: type }));
+      a.download = `file.${option}`;
+
+      a.click();
+      a.remove();
+    }
+  }
 
   showDeleteModal = deleteMessage => {
     this.setState({ showDeleteModal: true, deleteMessage });
@@ -250,6 +295,18 @@ class Topic extends Root {
         {selectedTab !== 'configs' && roles.topic && roles.topic['topic/data/insert'] && (
           <aside>
             <li className="aside-button">
+              {this.state.isAllTopicDataSelected && (
+                <div className="btn mr-2">
+                  <Dropdown>
+                    <Dropdown.Toggle>
+                      <strong>Download Format:</strong> ({this.state.downloadFormat})
+                    </Dropdown.Toggle>
+                    <Dropdown.Menu>
+                      <div>{this._renderDownloadFormat(this.state.isAllTopicDataSelected)}</div>
+                    </Dropdown.Menu>
+                  </Dropdown>
+                </div>
+              )}
               {this.canEmptyTopic() ? (
                 <div
                   onClick={() => {
