+++ client/src/containers/Topic/Topic/TopicData/TopicData.jsx
@@ -411,26 +411,8 @@ class TopicData extends Root {
     a.remove();
   }
 
-  _handleDownloadAll(option) {
-    let messages = this.state.messages;
-    if (!this.state.loading && this.props.isAllTopicDataSelected && option !== 'Select') {
-      let allData = [];
-      messages.map(tableData => {
-        allData.push(tableData.value);
-        allData.push('\n');
-      });
-      const a = document.createElement('a');
-      const type = 'text/' + option;
-      a.href = URL.createObjectURL(new Blob([allData], { type: type }));
-      a.download = `file.${option}`;
-
-      a.click();
-      a.remove();
-    }
-  }
-
   _handleCheckbox(e) {
-    this.props.onSelectAllCheckboxChange(e.target.checked);
+    this.props.onSelectAllCheckboxChange(e.target.checked, this.state.messages);
   }
 
   async _handleOnDateTimeFormatChanged(newDateTimeFormat) {
@@ -770,7 +752,7 @@ class TopicData extends Root {
       loading
     } = this.state;
 
-    let actions = [constants.TABLE_SHARE, constants.TABLE_COPY, constants.TABLE_DOWNLOAD_ALL];
+    let actions = [constants.TABLE_SHARE, constants.TABLE_COPY];
     if (canDeleteRecords) actions.push(constants.TABLE_DELETE);
     if (canDownload) actions.push(constants.TABLE_DOWNLOAD);
 
@@ -986,7 +968,9 @@ class TopicData extends Root {
                 type: 'checkbox',
                 expand: true,
                 cell: () => {
-                  return <input type="checkbox" checked={this.props.isAllTopicDataSelected} />;
+                  return (
+                    <input type="checkbox" checked={this.props.isAllTopicDataSelected} readOnly />
+                  );
                 },
                 readOnly: true
               },
@@ -1144,9 +1128,6 @@ class TopicData extends Root {
             }}
             onDownload={row => {
               this._handleDownload(row);
-            }}
-            onDownloadAll={option => {
-              this._handleDownloadAll(option);
             }}
             onCopy={row => {
               this._handleCopy(row);
